add clarifying comment and cast

+++ src/jumphash.c
@@ -37,5 +37,5 @@ int32_t jumphash(uint64_t key, int32_t num_buckets)
 		key = key * Linear_Congruential_Generator_64 + 1;
 		j = (b + 1) * ((double)(1LL << 31) / (double)((key >> 33) + 1));
 	}
-	return b;
+	return ((int32_t)b);
 }
